requirements.txt
streamlit
torch
librosa
ffmpeg-python
transformers